pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "experiment-builder"
description = "Facilitates and formalizes the experimentation process for computer vision."
authors = [{ name = "Andrew Root", email = "root.drew27@gmail.com" }]
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["numpy", "cupy-cuda12x", "pillow"]
readme = "README.md"

[tool.setuptools]
packages = ["expb"]

[tool.ruff]
line-length = 100
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
] # NOTE: ruff ignores files that are ignored by git
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true

[project.optional-dependencies]
formatting = ["ruff == 0.9.3"]
type-checking = ["mypy"]
dev = ["experiment-builder[formatting, type-checking]"]
